+++ scripts/macros/items/bootsOfElvenkind.js
@@ -1,6 +1,3 @@
-import {chris} from '../../helperFunctions.js';
-export async function bootsOfElvenkind(skillId, options) {
-    if (skillId != 'ste' || options.advantage) return;
-    let selection = await chris.dialog('Boots of Elvenkind', [['Yes', true], ['No', false]], 'Does this check rely on moving silently?');
-    if (selection) options.advantage = true;
+export function bootsOfElvenkind(skillId, options) {
+    return skillId != 'ste' ? false : {'label': 'This check relies on moving silently.', 'type': 'advantage'};
 }