+++ scripts/macros/items/eyesOfMinuteSeeing.js
@@ -1,6 +1,3 @@
-import {chris} from '../../helperFunctions.js';
-export async function eyesOfMinuteSeeing(skillId, options) {
-    if (skillId != 'inv' || options.advantage) return;
-    let selection = await chris.dialog('Eyes of Minute Seeing', [['Yes', true], ['No', false]], 'Does this check rely on sight while<br>searching an area or studying an object?');
-    if (selection) options.advantage = true;
+export function eyesOfMinuteSeeing(skillId, options) {
+    return skillId != 'inv' ? false : {'label': 'This check relies on searching or studying and object.', 'type': 'advantage'};
 }