+++ scripts/macros/items/eyesOfTheEagle.js
@@ -1,6 +1,3 @@
-import {chris} from '../../helperFunctions.js';
-export async function eyesOfTheEagle(skillId, options) {
-    if (skillId != 'prc' || options.advantage) return;
-    let selection = await chris.dialog('Eyes of the Eagle', [['Yes', true], ['No', false]], 'Does this check rely on sight?');
-    if (selection) options.advantage = true;
+export function eyesOfTheEagle(skillId, options) {
+    return skillId != 'prc' ? false : {'label': 'This check relies on sight.', 'type': 'advantage'};
 }