+++ scripts/macros/monsterFeatures/deepGnome/stoneCamouflage.js
@@ -1,6 +1,3 @@
-import {chris} from '../../../helperFunctions.js';
-export async function stoneCamouflage(skillId, options) {
-    if (skillId != 'ste' || options.advantage) return;
-    let selection = await chris.dialog('Stone Camouflage', [['Yes', true], ['No', false]], 'Hiding in rocky terrain?');
-    if (selection) options.advantage = true;
+export function stoneCamouflage(skillId, options) {
+    return skillId != 'ste' ? false : {'label': 'Hiding in rocky terrain?', 'type': 'advantage'};
 }