+++ scripts/macros/raceFeatures/minotaur/labyrinthineRecall.js
@@ -1,0 +1,3 @@
+export function labyrinthineRecall(skillId, options) {
+    return skillId != 'sur' ? false : {'label': 'This check is to navigate or track.', 'type': 'advantage'};
+}